Add back navigation props for cloud
