+++ src/lib/holocene/navigation/full-nav.svelte
@@ -7,7 +7,6 @@
 </script>
 
 <script lang="ts">
-  import { page } from '$app/stores';
   import Icon from '$holocene/icon/icon.svelte';
   import type { SvelteComponent } from 'svelte';
   import FeatureTag from '$lib/holocene/feature-tag.svelte';
@@ -24,37 +23,18 @@
   import { afterNavigate } from '$app/navigation';
   import { viewFeature } from '$lib/stores/new-feature-tags';
   import FeatureGuard from '$lib/components/feature-guard.svelte';
-  import { fetchNamespaces } from '$lib/services/namespaces-service';
-  import { lastUsedNamespace } from '$lib/stores/namespaces';
-  import {
-    routeForArchivalWorkfows,
-    routeForWorkflows,
-    routeForSchedules,
-    routeForNamespaces,
-  } from '$lib/utilities/route-for';
-  import type { ListNamespacesResponse } from '$types';
 
+  import type { DescribeNamespaceResponse as Namespace } from '$types';
+
+  export let isCloud = false;
+  export let activeNamespace: Promise<Namespace>;
+  export let getNamespaceList: () => Promise<NamespaceItem[]>;
+  export let linkList: Partial<Record<string, string>>;
   export let user: Promise<User> | undefined;
   export let logout: () => void;
   export let extras: ExtraIcon[] | null = null;
 
-  const { isCloud } = $page.stuff.settings.runtimeEnvironment;
-  const namespacesPromise: Promise<ListNamespacesResponse> = fetchNamespaces(
-    $page.stuff.settings,
-  );
   let showProfilePic = true;
-
-  $: activeNamespaceName = $page.params?.namespace ?? $lastUsedNamespace;
-  $: linkList = {
-    home: routeForWorkflows({ namespace: activeNamespaceName }),
-    archive: routeForArchivalWorkfows({ namespace: activeNamespaceName }),
-    namespaces: routeForNamespaces(),
-    schedules: routeForSchedules({ namespace: activeNamespaceName }),
-    workflows: routeForWorkflows({ namespace: activeNamespaceName }),
-    feedback:
-      $page.stuff?.settings?.feedbackURL ||
-      'https://github.com/temporalio/ui/issues/new/choose',
-  };
 
   function fixImage() {
     showProfilePic = false;
@@ -78,12 +58,9 @@
       <div class="nav-title">Workflows</div>
     </NavRow>
     <IsCloudGuard {isCloud}>
-      {#await namespacesPromise then namespacesResponse}
-        {@const activeNamespace = namespacesResponse.namespaces.find(
-          (namespace) => namespace?.namespaceInfo?.name === activeNamespaceName,
-        )}
+      {#await activeNamespace then namespace}
         <FeatureGuard
-          enabled={Boolean(activeNamespace?.namespaceInfo?.supportsSchedules)}
+          enabled={Boolean(namespace?.namespaceInfo?.supportsSchedules)}
         >
           <NavRow
             link={linkList.schedules}
@@ -212,6 +189,7 @@
     >
       {#if $namespaceSelectorOpen}
         <NamespaceList
+          {getNamespaceList}
           on:closeNamespaceList={(event) => {
             $namespaceSelectorOpen = false;
           }}
