+++ src/lib/stores/namespaces.ts
@@ -1,3 +1,3 @@
 import { persistStore } from './persist-store';
 
-export const lastUsedNamespace = persistStore('lastNamespace', 'default');
+export const lastUsedNamespace = persistStore<string>('lastNamespace', 'default');
